+++ CHANGELOG.md
@@ -30,6 +30,9 @@ Format is based on [Keep a Changelog 1.1.0](https://keepachangelog.com/en/1.1.0/
 - **React Native:** Logs are now visible in Xcode on iOS.
 - **Swift:** Logs are now visible in Xcode on iOS.
 
+### Removed
+- **Flutter:** Removed `ToolCallExtension` and `ToolCall.argumentsJson` as `ToolCall` is no longer opaque.
+
 ## [Python v2.0.0, Flutter v3.0.0, Godot v10.0.0, Kotlin v3.0.0, React Native v3.0.0, Swift v3.0.0] - 2026-08-20
 
 ### Added